move kelas and mahasiswa to model
