+++ rekapnilai/src/main/java/com/k4/rekapnilai/model/Kelas.java
@@ -1,7 +1,8 @@
-package com.k4.rekapnilai;
+package com.k4.rekapnilai.model;
 
 public class Kelas{
-	private String NIM,kelas;
+	private String NIM, kelas;
+	
 	public Kelas(String namaMhs, String NIM, String kelas) {
 		
 		this.NIM = NIM;
